requirements.txt
tabgenie
transformers == 4.25.1
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113; sys_platform == "linux"
evaluate
click
sacrebleu
jupyter
nltk
python-dateutil
dateparser==1.1.8
gdown
neptune-client==0.16.13
pytimeparse==1.1.8
unidecode==1.3.6
statsmodels==0.14.0